modified:   .config/waybar/scripts/vpn-watch.sh 	modified:   .config/waybar/scripts/vpn.sh

+++ .config/waybar/scripts/vpn-watch.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env bash
 
+lastline=
+
 while read -r line; do
+  if [[ $lastline == $line ]] \
+  || echo $line | grep Deleted &>/dev/null; then
     pkill -RTMIN+8 waybar
+  fi
+  if echo $line | grep mullvad &>/dev/null; then
+    lastline=$(echo $line)
+  fi
 done < <(ip monitor link)
